requirements.txt
streamlit
streamlit_chat
langchain
python-dotenv
langchain_community
langchain_openai
pypdf
faiss-cpu
langchain_groq
-e .